requirements.txt
langchain
langgraph
langchain-openai
faiss-cpu
tiktoken
python-dotenv
langchain-core
langchain-community
langchain-google-genai